fix/refractor: fix bug with adding to room after plaer disconect. refractor closeConection and index.ts
