+++ backend/index.ts
@@ -1,7 +1,18 @@
-import { App } from './src/app';
-import { downApp } from './src/helpers/shotDownApp';
+import fs from 'fs';
+import path from 'path';
+import http from 'http';
 
-const app = new App();
-app.serverRun();
-
-downApp(app);
+export const httpServer = http.createServer(function (req, res) {
+  const __dirname = path.resolve(path.dirname(''));
+  const file_path =
+    __dirname + (req.url === '/' ? '/front/index.html' : '/front' + req.url);
+  fs.readFile(file_path, function (err, data) {
+    if (err) {
+      res.writeHead(404);
+      res.end(JSON.stringify(err));
+      return;
+    }
+    res.writeHead(200);
+    res.end(data);
+  });
+});
