+++ backend/src/app.ts
@@ -2,11 +2,7 @@ import { Server, WebSocketServer } from 'ws';
 import { print } from './helpers/print';
 import { getMessage } from './messages/getMessage';
 import { WebSocketId } from './types/webSocket';
-import { returnCurrentPlayer } from './helpers/returnCurrent';
-import { DB } from './dataBase/dataBase';
-import { updateWiners } from './helpers/updateWiners';
-import { AnswerFinishData, EnumTypes } from './types/iServerMsg';
-import { sendMessage } from './helpers/sendMessage';
+import { closeConection } from './helpers/closeConection';
 
 const PORT = 3000;
 
@@ -31,29 +27,7 @@ export class App {
 
       ws.on('close', () => {
         print('Conection fallen', 'red');
-        const [game] = DB.games.filter((game) =>
-          game.players.some((player) => player.id === ws.id),
-        );
-        if (!game) return;
-
-        const [player] = game.players.filter((player) => player.id !== ws.id);
-
-        const winner = returnCurrentPlayer(player.id);
-        winner.wins += 1;
-        const data: AnswerFinishData = {
-          winPlayer: winner.index,
-        };
-        try {
-          this.server.clients.forEach((client: WebSocketId) => {
-            if (client.id === winner.id)
-              sendMessage(client, EnumTypes.finish, data);
-          });
-        } catch {
-          print('Some went wrong', 'red');
-        }
-        updateWiners(this.server);
-        DB.deleteGame(game.idGame);
-        DB.deleteRoom(game.idGame);
+        closeConection(ws, this.server);
       });
     });
   }
