+++ package.json
@@ -2,10 +2,10 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
+    "main": "index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "ts-node ./index.ts",
+        "start:dev": "nodemon ./index.ts",
         "server:dev": "nodemon backend/index.ts",
 				"server": "ts-node backend/index.ts",
         "test": "echo \"Error: no test specified\" && exit 1",
